Moves example to its own files
