+++ server/mongodb/actions/Example.ts
@@ -1,0 +1,6 @@
+import mongoDB from '../index'
+
+export async function testFunction2 (): Promise<String> {
+  await mongoDB()
+  return 'Example string'
+}
